paths

+++ pyth/pydbutil/main.py
@@ -48,9 +48,10 @@ def main(argv=None):
                          for k, v in ms.matches.iteritems() if len(v) > 0]    
             print ms.summary_files()
             #print ms.summary_details()
-            FileUtils.ensure_dir(ConfigMgr.WORK_DIR)
-            shutil.rmtree(ConfigMgr.WORK_DIR)
             
+            if os.path.exists(ConfigMgr.WORK_DIR) :
+                shutil.rmtree(ConfigMgr.WORK_DIR)
+                        
             FileUtils.copy_files(sourcepaths, targpaths, DO_ASK)
             print len(sourcepaths), 'file(s) copied to work directory', ConfigMgr.WORK_DIR
         if DO_MOD:
